Move SlicerOpenLIFUSolution to standalone python module (#95)
